% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_site_board.R
\name{flex_site_board}
\alias{flex_site_board}
\title{Flexible Dashboards Split in Sub-Pages}
\usage{
flex_site_board(...)
}
\arguments{
\item{...}{see \code{flexdashbaord}

Hint: Needs all global JavaScript code before the first header.}
}
\value{
a new format for \code{rmarkdown}
}
\description{
Flexible Dashboards Split in Sub-Pages
}
