% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_input_formats}
\alias{pandoc_input_formats}
\title{Give all Input Formats Supported by \code{Pandoc}}
\usage{
pandoc_input_formats(use_cache = TRUE)
}
\arguments{
\item{use_cache}{don't call \code{Pandoc} again}
}
\value{
all supported \code{Pandoc} input formats
}
\description{
Give all Input Formats Supported by \code{Pandoc}
}
