% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc_output_formats}
\alias{pandoc_output_formats}
\title{Give all Output Formats Supported by \code{Pandoc}}
\usage{
pandoc_output_formats(use_cache = TRUE)
}
\arguments{
\item{use_cache}{don't call \code{Pandoc} again}
}
\value{
all supported \code{Pandoc} output formats
}
\description{
Give all Output Formats Supported by \code{Pandoc}
}
