% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc_read.R
\name{pandoc_read}
\alias{pandoc_read}
\title{Read a \code{Pandoc} File from some Supported Format}
\usage{
pandoc_read(file, text = NULL, from = pandoc_input_formats())
}
\arguments{
\item{file}{file to read}

\item{text}{string to use instead of file if set}

\item{from}{format of the file}
}
\value{
the \code{Pandoc} object
}
\description{
Read a \code{Pandoc} File from some Supported Format
}
