% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc_write.R
\name{pandoc_write}
\alias{pandoc_write}
\title{Write a \code{Pandoc} File}
\usage{
pandoc_write(pandoc, to = pandoc_output_formats())
}
\arguments{
\item{pandoc}{internal \code{Pandoc} representation}

\item{to}{output format (supported by \code{Pandoc})}
}
\value{
the converted text
}
\description{
Write a \code{Pandoc} File
}
